adding lang-utils + refactoring the code using the new functions
