+++ src/components/Footer.astro
@@ -2,7 +2,7 @@
 import GithubIcon from "@/icons/footer/GithubIcon.astro";
 import InstagramIcon from "@/icons/footer/InstagramIcon.astro";
 import TwitterIcon from "@/icons/footer/TwitterIcon.astro";
-import { getTranslateLabels } from "../lib/lang-utils";
+import { getTranslateLabels } from "../i18n/utils";
 
 const translateLabels = getTranslateLabels(Astro.currentLocale);
