+++ src/sections/AboutSection.astro
@@ -1,6 +1,6 @@
 ---
 import { aboutParagraphs } from "@/data/about.ts";
-import { getTranslateLabels, currentLang } from "../lib/lang-utils";
+import { getTranslateLabels, currentLang } from "../i18n/utils";
 
 const translateLabels = getTranslateLabels(Astro.currentLocale);
 const paragraphs = aboutParagraphs[currentLang(Astro.currentLocale) as keyof typeof aboutParagraphs];
