+++ src/sections/Error.astro
@@ -1,6 +1,6 @@
 ---
 import LogoCaawards from '@/icons/LogoCaawards.astro';
-import { getTranslateLabels } from "../lib/lang-utils";
+import { getTranslateLabels } from "../i18n/utils";
 
 const translateLabels = getTranslateLabels(Astro.currentLocale);
 ---
